refactor: move getLegacyImporterList in a separated file
